Update G. Summation from 1 to N.java

+++ Sheet #1 (Data type - Conditions)/G. Summation from 1 to N.java
@@ -3,7 +3,7 @@ import java.util.*;
 public class Main {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
-        int n = sc.nextInt();
+        long n = sc.nextLong();
         System.out.println(n*(n+1)/2);
     }
 }
